Add analytics & Detach component `ProfileBanner`
